Adding HTML/JavaScript Implementation
pythy.html is a web page, which uses HTML5 canvas and some
Javascript to produce a couple figures illustrating this principle
and animation showing how the proportions work out over a broad
range of base x height proportions.

The colors are adjusted from the Python version.

There are no external dependencies of this HTML page ... it runs
entirely in the browser and uses no JQuery nor other JS libraries.

It could use additional CSS styling and the proportions of figure
size to the canvas dimensions are a bit off.